fix: Confirmation Discard modal code refactoring
