+++ apps/ai-dial-admin/src/components/Assets/Prompts/View/View.tsx
@@ -25,7 +25,6 @@ import { getUrnForEntity } from '@/src/utils/open-in-new-tab';
 import { EntityViewTab, getTabsForAsset } from '@/src/utils/tabs/utils';
 import { addTrailingSlash } from '@/src/utils/url';
 import TabsContent from './TabsContent';
-import DiscardModal from '@/src/components//EntityView/Modals/Discard/Discard';
 
 interface Props {
   originalPrompt: DialPrompt;
@@ -45,7 +44,6 @@ const PromptView: FC<Props> = ({ originalPrompt, etag, prompts }) => {
   const [selectedPrompt, setSelectedPrompt] = useState(structuredClone(originalPrompt));
   const [isChanged, setIsChanged] = useState(false);
   const [isEditorEnabled, setIsEditorEnabled] = useState(false);
-  const [isDiscardModalOpen, setIsDiscardModalOpen] = useState(false);
 
   const [addedVersions, setAddedVersions] = useState<string[]>([]);
 
@@ -122,22 +120,13 @@ const PromptView: FC<Props> = ({ originalPrompt, etag, prompts }) => {
     [selectedPrompt, originalPrompt, etag, showNotification, t, router, fetchFiles],
   );
 
-  const onTryToDiscard = useCallback(() => {
-    setIsDiscardModalOpen(true);
-  }, []);
-
-  const onDiscardModalConfirm = useCallback(() => {
-    onDiscard();
-    setIsDiscardModalOpen(false);
-  }, [onDiscard]);
-
   return (
     <div className="flex flex-col flex-1 min-h-0 w-full bg-layer-2 rounded p-4 pb-14 lg:pb-4 relative">
       <AssetHeader
         view={ApplicationRoute.Prompts}
         entity={selectedPrompt}
         isChanged={isChanged}
-        onDiscard={onTryToDiscard}
+        onDiscard={onDiscard}
         onSave={onSave}
         tabs={tabs}
         assets={prompts}
@@ -155,16 +144,7 @@ const PromptView: FC<Props> = ({ originalPrompt, etag, prompts }) => {
         {isEditorEnabled ? (
           <EntityJsonEditor entity={selectedPrompt} setSelectedEntity={setSelectedPrompt} setIsChanged={setIsChanged} />
         ) : (
-          <>
-            <TabsContent activeTab={activeTab} onChangePrompt={setSelectedPrompt} selectedPrompt={selectedPrompt} />
-            {isDiscardModalOpen && (
-              <DiscardModal
-                onConfirm={onDiscardModalConfirm}
-                onClose={() => setIsDiscardModalOpen(false)}
-                onCancel={() => setIsDiscardModalOpen(false)}
-              />
-            )}
-          </>
+          <TabsContent activeTab={activeTab} onChangePrompt={setSelectedPrompt} selectedPrompt={selectedPrompt} />
         )}
       </div>
     </div>
