+++ apps/ai-dial-admin/src/components/EntityHeaderControls/Wrappers/AssetButtonsWrapper.tsx
@@ -27,6 +27,7 @@ import { useI18n } from '@/src/locales/client';
 import { AssetWithVersion } from '@/src/models/dial/deployment-asset';
 import AssetChangedEntityButtons from '../Buttons/AssetChangedEntityButtons';
 import { SimpleButtonsWrapperProps } from './SimpleButtonsWrapper';
+import DiscardModal from '@/src/components//EntityView/Modals/Discard/Discard';
 
 export interface AssetButtonsWrapperProps extends Omit<SimpleButtonsWrapperProps<AssetWithVersion>, 'onSave'> {
   assets?: AssetWithVersion[] | null;
@@ -66,6 +67,7 @@ const AssetButtonsWrapper: FC<AssetButtonsWrapperProps> = ({
   const isMobile = useIsMobileScreen();
 
   const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isDiscardModalOpen, setIsDiscardModalOpen] = useState(false);
   const [containerClassName, setContainerClassName] = useState(SELECT_ENTITY_HEADER_CLASS);
   const [buttonsClassName, setButtonsClassName] = useState('');
 
@@ -90,6 +92,15 @@ const AssetButtonsWrapper: FC<AssetButtonsWrapperProps> = ({
     onDiscard?.();
   }, [dispatch, onDiscard]);
 
+  const onTryToDiscard = useCallback(() => {
+    setIsDiscardModalOpen(true);
+  }, []);
+
+  const onDiscardModalConfirm = useCallback(() => {
+    onStartDiscard();
+    setIsDiscardModalOpen(false);
+  }, [onStartDiscard]);
+
   const onTryToSave = useCallback(
     (version?: string) => {
       if (jsonErrors?.length) {
@@ -111,7 +122,7 @@ const AssetButtonsWrapper: FC<AssetButtonsWrapperProps> = ({
           <AssetChangedEntityButtons
             version={entity.version}
             existingVersions={existingVersions}
-            onDiscard={onStartDiscard}
+            onDiscard={onTryToDiscard}
             isEditorEnabled={isEditorEnabled}
             onSave={onTryToSave}
           />
@@ -155,6 +166,13 @@ const AssetButtonsWrapper: FC<AssetButtonsWrapperProps> = ({
           />,
           document.body,
         )}
+      {isDiscardModalOpen && (
+        <DiscardModal
+          onConfirm={onDiscardModalConfirm}
+          onClose={() => setIsDiscardModalOpen(false)}
+          onCancel={() => setIsDiscardModalOpen(false)}
+        />
+      )}
     </>
   );
 };
