+++ apps/ai-dial-admin/src/components/Keys/View/View.tsx
@@ -27,7 +27,6 @@ import { EntityViewTab, getKeyTabs } from '@/src/utils/tabs/utils';
 import KeyRotateModal from '../Modals/KeyRotateModal';
 import TabsContent from './TabsContent';
 import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
-import DiscardModal from '@/src/components//EntityView/Modals/Discard/Discard';
 
 interface Props {
   originalKey: DialKey;
@@ -50,7 +49,6 @@ const KeyView: FC<Props> = ({ originalKey, etag, ...props }) => {
   const [selectedKey, setSelectedKey] = useState(cloneDeep(originalKey));
   const [isChanged, setIsChanged] = useState(false);
   const [isEditorEnabled, setIsEditorEnabled] = useState(false);
-  const [isDiscardModalOpen, setIsDiscardModalOpen] = useState(false);
 
   const [selectedFormat, setSelectedFormat] = useState(ExportFormat.ADMIN);
   const [coreKey, setCoreKey] = useState<DialKey | null>(null);
@@ -145,15 +143,6 @@ const KeyView: FC<Props> = ({ originalKey, etag, ...props }) => {
     }
   }, [selectedKey, selectedFormat, onSaveKey]);
 
-  const onTryToDiscard = useCallback(() => {
-    setIsDiscardModalOpen(true);
-  }, []);
-
-  const onDiscardModalConfirm = useCallback(() => {
-    onDiscard();
-    setIsDiscardModalOpen(false);
-  }, [onDiscard]);
-
   return (
     <>
       <div className="flex flex-col flex-1 min-h-0 w-full bg-layer-2 rounded p-4 pb-14 lg:pb-4 relative">
@@ -161,7 +150,7 @@ const KeyView: FC<Props> = ({ originalKey, etag, ...props }) => {
           view={ApplicationRoute.Keys}
           entity={selectedKey}
           isChanged={isChanged}
-          onDiscard={onTryToDiscard}
+          onDiscard={onDiscard}
           onSave={onTryToSaveKey}
           tabs={tabs}
           jsonConfiguration={jsonConfiguration}
@@ -216,14 +205,6 @@ const KeyView: FC<Props> = ({ originalKey, etag, ...props }) => {
           />,
           document.body,
         )}
-
-      {isDiscardModalOpen && (
-        <DiscardModal
-          onConfirm={onDiscardModalConfirm}
-          onClose={() => setIsDiscardModalOpen(false)}
-          onCancel={() => setIsDiscardModalOpen(false)}
-        />
-      )}
     </>
   );
 };
