+++ apps/ai-dial-admin/src/components/Routes/View/View.tsx
@@ -9,8 +9,6 @@ import { getCoreRoute, removeRoute, updateCoreRoute, updateRoute } from '@/src/a
 import { JsonConfiguration } from '@/src/components/EntityHeaderControls/models';
 import SimpleEntityHeader from '@/src/components/EntityHeaderControls/SimpleHeader';
 import EntityJsonEditor from '@/src/components/EntityTabs/JsonEditor/JsonEditor';
-import { ModalType } from '@/src/components/EntityView/Modals/constants';
-import EntityViewModals from '@/src/components/EntityView/Modals/EntityViewModals';
 import { useNotification } from '@/src/context/NotificationContext';
 import { useSaveValidationContext, ValidationActionType } from '@/src/context/SaveValidationContext';
 import { useProtectedRequest } from '@/src/hooks/use-protected-request';
@@ -44,9 +42,6 @@ const RouteView: FC<Props> = ({ originalRoute, etag, names, roles }) => {
   const [selectedRoute, setSelectedRoute] = useState(cloneDeep(originalRoute));
   const [isChanged, setIsChanged] = useState(false);
   const [isEditorEnabled, setIsEditorEnabled] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [modalType, setModalType] = useState<ModalType>();
-
   const [selectedFormat, setSelectedFormat] = useState(ExportFormat.ADMIN);
   const [coreRoute, setCoreRoute] = useState<DialRoute | null>(null);
 
@@ -116,46 +111,13 @@ const RouteView: FC<Props> = ({ originalRoute, etag, names, roles }) => {
     });
   }, [selectedFormat, selectedRoute, originalRoute.name, etag, dispatch, showNotification, t, router]);
 
-  const onModalClose = useCallback(() => {
-    setIsModalOpen(false);
-    setModalType(void 0);
-  }, []);
-
-  const onModalOpen = useCallback((modalType: ModalType) => {
-    setModalType(modalType);
-    setIsModalOpen(true);
-  }, []);
-
-  const onModalCancel = useCallback(
-    (type: ModalType) => {
-      if (type === ModalType.discard) {
-        onModalClose();
-      }
-    },
-    [onModalClose],
-  );
-
-  const onModalConfirm = useCallback(
-    (type: ModalType) => {
-      if (type === ModalType.discard) {
-        onDiscard();
-        onModalClose();
-      }
-    },
-    [onDiscard, onModalClose],
-  );
-
-  const onTryToDiscard = useCallback(() => {
-    onModalOpen(ModalType.discard);
-  }, [onModalOpen]);
-
   return (
     <div className="flex flex-col flex-1 min-h-0 w-full bg-layer-2 rounded p-4 pb-14 lg:pb-4 relative">
       <SimpleEntityHeader
         view={ApplicationRoute.Routes}
         entity={selectedRoute}
         isChanged={isChanged}
-        onDiscard={onTryToDiscard}
+        onDiscard={onDiscard}
         onSave={onSaveRoute}
         tabs={tabs}
         jsonConfiguration={jsonConfiguration}
@@ -177,14 +139,6 @@ const RouteView: FC<Props> = ({ originalRoute, etag, names, roles }) => {
           />
         )}
       </div>
-
-      <EntityViewModals
-        isModalOpen={isModalOpen}
-        modalType={modalType}
-        handleConfirm={onModalConfirm}
-        handleClose={onModalClose}
-        handleCancel={onModalCancel}
-      />
     </div>
   );
 };
